requirements.txt
geopandas>=0.14.0
folium>=0.15.0
mapclassify>=2.6.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
jupyterlab>=4.0.0
